third_party/WebKit/LayoutTests/platform/mac-mac10.11/fast/forms/calendar-picker/calendar-picker-should-not-change-datetimelocal-time-expected.txt

Test for crbug.com/271451. The calendar picker should not change the time part of datetime-local field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getUserAgentShadowTextContent(input) is "1999-07-31 12:59"
PASS getUserAgentShadowTextContent(input) is "1999-07-31 12:59"
FAIL highlightedValue() should be 1999-08-01. Was 1999-07-21.
FAIL getUserAgentShadowTextContent(input) should be 1999-08-01 12:59. Was 1999-07-21 12:59.
PASS successfullyParsed is true

TEST COMPLETE
